+++ docs/changes.md
@@ -1,5 +1,9 @@
 # 最新动态
 
+* 2019/12/17
+  * 完善dialog\_confirm/dialog\_info等函数。
+  * 增加widget\_set\_focusable函数。
+
 * 2019/12/13
   * 修改窗口切到FOREGROUND把焦点设回去会导致如果当前页面有滚动控件时，页面会滚动的问题（感谢大恒提供补丁）。
   * 修改widget在point_down少发了个leave事件的问题（感谢大恒提供补丁）。